backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
pytest
openai
alembic

